Add schedule-latency crate feature
